refactor: 상수 파일 경로를 'constant'에서 'constants'로 변경
