+++ frontend/src/features/meeting/constants/conditionCard.ts
